Use custom Gemfile with example app.
Without using a custom Gemfile, the main `rspec-rails` repo bundle is
used. Due to how various Rails autoload issues work with bundler, this
caused `RSpec::Support` to be automatically loaded by the rails
generator. The root cause of this is because `rspec-support` and
`rspec-core` are explicitly included in the Gemfile.

This differs in behavior from when a normal Rail's app only specifies
`rspec-rails` in the Gemfile. Setup an explicit Gemfile to resolve the
load difference issues.